docker/whisper-diarize/requirements.txt
faster-whisper>=1.0.0
pyannote.audio>=3.1.1
torch>=2.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydub>=0.25.1
